+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Version 1.78 (not yet released)
 ------------
 
+* support note date on 'Create Note'  by !53
+
 Version 1.77
 ------------